+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Please re-run **killtracker_load_eve** to get all the newly added types.
 ### Changed
 
 - Restructured tasks to improve scalability, performance and resilience [#20](https://gitlab.com/ErikKalkoken/aa-killtracker/-/issues/20)
+- Killtracker will no longer start when ESI is offline
 - Remove support for Django 2.1 & 3.0 [#17](https://gitlab.com/ErikKalkoken/aa-killtracker/-/issues/17)
 
 ## [0.2.6] - 2020-12-09